Files From Getting to Know Android SDK
Here are some of the files I studied while attempting to create a basic Hello World program in Android SDK. I modified strings.xml and activity_main.xml. There are a lot of components created by default for an Android app, and so these four files of themselves won't really do anything. I'm working on running the project in an emulator. Will upload a screenshot to the related issue the when I have it.